pyproject.toml
[project]
name = "yellow-docs-mcp"
version = "0.1.0"
description = "MCP server for Yellow Network documentation - search, browse, and retrieve docs with AI"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Gianluca Panza", email = "panzaguardattigianluca@gmail.com" },
]
readme = "README.md"
keywords = ["mcp", "yellow-network", "documentation", "ai", "state-channels", "nitrolite"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "mcp[cli]>=1.0",
    "sentence-transformers>=2.2",
    "rank-bm25>=0.2",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "gitpython>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.urls]
Homepage = "https://github.com/panzagianluca/yellow-docs-mcp"
Repository = "https://github.com/panzagianluca/yellow-docs-mcp"
Documentation = "https://docs.yellow.org"
Issues = "https://github.com/panzagianluca/yellow-docs-mcp/issues"

[project.scripts]
yellow-docs-mcp = "yellow_docs_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = [
    "live: runs tests that require internet access and live upstream repositories",
]
